Fixed sound effects
Removed the queue_free() from the Enemy, which was the reason the audio clip was playing only 1 frame - because you were deleting the whole Enemy Entity right at the time of death. Probably you forgot to remove this line when you implemented the DestroyedComponent.
I also fixed your HealthComponent to not emit the "death" signal more than once, because that triggered the sound effect to play multiple times if you hit the Enemy after he was already at 0 health.